Provide more useful indices for lookups on database (#209)
On a search lookup from mongodatabase there can be only one index applied for search. this index could be compound.
As the lookups are all time including exchange and symbol + other fields its better to store them all in a compount index, which can then be used for a lookup.
Check the explain plan of mongodb for this.